moved some methods from roomcontroller to bookingcontroller
