+++ Hotel/src/Application/controllers/BookingController.java
@@ -1,7 +1,11 @@
 package Application.controllers;
 
+import Application.Interfaces.SaveBooking;
 import Application.models.Booking;
+import Application.models.Guest;
 import Application.models.User;
+import Application.utils.CSVWriter;
+import Application.utils.DBSaver;
 
 import java.text.SimpleDateFormat;
 import java.time.LocalDateTime;
@@ -106,5 +110,51 @@ public class BookingController {
         SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yyyy");
         Date endDate = formatter.parse(date);
         return endDate;
+    }
+    public void createAndSaveBooking(Booking booking){
+        User guest = login();
+        setUserDetails(booking, guest);
+        saveBooking(booking,"csv");
+    }
+    private User login(){
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("User name: ");
+        String userName = scanner.nextLine();
+        System.out.println("Enter password: ");
+        String enteredPassword = scanner.nextLine();
+        User user = new User(userName, enteredPassword);
+        return user;
+    }
+    private void setUserDetails(Booking booking, User user){
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("First name: ");
+        String firstName = scanner.nextLine();
+        System.out.println("Last name: ");
+        String lastName = scanner.nextLine();
+        System.out.println("Address: ");
+        String address = scanner.nextLine();
+        System.out.println("houseNumber: ");
+        String houseNumber = scanner.nextLine();
+        System.out.println("Postcode: ");
+        String postcode = scanner.nextLine();
+        System.out.println("City: ");
+        String city = scanner.nextLine();
+        System.out.println("country: ");
+        String country = scanner.nextLine();
+        System.out.println("Email address: ");
+        String emailAddress = scanner.nextLine();
+        System.out.println("Telephone number: ");
+        String phoneNumber = scanner.nextLine();
+        Guest guest = new Guest(user, phoneNumber, address, houseNumber, postcode, city,country, emailAddress);
+        booking.setHeadBooker(guest);
+    }
+    private void saveBooking(Booking booking, String type){
+        SaveBooking saveInstance;
+        if(type.equals("csv")) {
+            saveInstance = new CSVWriter();
+        } else{
+            saveInstance = new DBSaver();
+        }
+        saveInstance.saveBooking(booking);
     }
 }
